Revert "Merge pull request #15901 from metasfresh/scary_plum_uat_merge_from_release"
This reverts commit 02fab34a7032f38d687ee1b4ab2160e2638195f2, reversing
changes made to 9b6bbccf4f7ab01eb690762a60b0bd22b3f08cb2.

+++ backend/de.metas.ui.web.base/src/main/java/de/metas/ui/web/kpi/data/KPIDataValue.java
@@ -134,7 +134,7 @@ public class KPIDataValue
 					}
 					else if (value instanceof Number)
 					{
-						final BigDecimal bd = BigDecimal.valueOf(((Number)value).doubleValue());
+						final BigDecimal bd = BigDecimal.valueOf(((Number)value).intValue());
 						return roundToPrecision(bd);
 					}
 					else
